Moved grouping tests to astro.tests.test_data.py
Since the tests were changed from using the UI
(e.g., "ui.load_pha()") to object-oriented ("data = read_pha()"),
the tests were moved out of test_astro_ui.py.
